pyproject.toml
[project]
name = "training-ground"
version = "0.1.0"
description = "Simple tool for training AI models."
readme = "README.md"
authors = [
    { name = "Tim Wedde", email = "tim@agrowizard.com" }
]
requires-python = ">=3.13"
dependencies = [
    "rfdetr>=1.5.2",
    "typer>=0.24.1",
    "roboflow>=1.2.16",
    "questionary>=2.1.1",
    "pycocotools>=2.0.11",
    "faster-coco-eval>=1.7.2",
    "pytorch-lightning>=2.6.1",
    "orjson>=3.11.7",
    "plotly>=6.3.1",
    "httpx>=0.27.0",
]

[project.scripts]
training-ground = "training_ground.main:app"

[build-system]
requires = ["uv_build>=0.10.11,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pyinstrument>=5.1.2",
]
